collecting users - rename package without underscores
